app/requirements.txt
fastapi==0.94.1
uvicorn==0.15.0
pydantic==1.10.2
mysql-connector-python==8.0.26